requirements.txt
discord.py[voice]
wavelink==1.3.5
discord.py-pagination==0.0.5
pytube==12.1.3